# backend/src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Comment System API",
    "version": "1.0.0",
    "description": "A comprehensive comment system API with authentication, comments, likes, and real-time updates via WebSocket",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token. Format: Bearer {token}"
      }
    },
    "schemas": {
...
      "CreateCommentRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "description": "Comment content",
            "example": "This is a great post!"
          },
          "parentId": {
            "type": "string",
            "description": "Parent comment ID (for replies)",
            "example": "clm1234567890abc"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Comment unique identifier",
            "example": "clm1234567890abc"
          },
          "content": {
            "type": "string",
            "description": "Comment content",
            "example": "This is a great post!"
          },
          "authorId": {
            "type": "string",
            "description": "Author user ID",
            "example": "clm1234567890abc"
          },
          "parentId": {
            "type": "string",
            "description": "Parent comment ID (for replies)",
            "example": "clm1234567890abc"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Comment creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Comment last update date",
            "example": "2024-01-01T00:00:00.000Z"
          }
        }
      },
      "CommentWithDetails": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Comment"
          },
          {
            "type": "object",
            "properties": {
              "author": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              },
              "_count": {
                "type": "object",
                "properties": {
                  "replies": {
                    "type": "number"
                  },
                  "likes": {
                    "type": "number"
                  }
                }
              },
              "likesCount": {
                "type": "number",
                "description": "Number of likes"
              },
              "dislikesCount": {
                "type": "number",
                "description": "Number of dislikes"
              }
            }
          }
        ]
      },
...
      "PaginationMetadata": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 10
          },
          "total": {
            "type": "number",
            "description": "Total number of items",
            "example": 100
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 10
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "error": {
            "type": "string",
            "description": "Error code",
            "example": "ERROR_CODE"
          },
          "errors": {
            "type": "array",
            "description": "Validation errors (if applicable)",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field name",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "description": "Error message",
                  "example": "Invalid email format"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/comments": {
      "post": {
        "summary": "Create a new comment",
        "description": "Create a new comment or reply to an existing comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CommentWithDetails"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Parent comment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get comments with pagination and sorting",
        "description": "Retrieve a paginated list of comments with optional filtering and sorting",
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "content",
                "likesCount",
                "dislikesCount",
                "repliesCount"
              ]
            },
            "description": "Field to sort by (comma-separated for multiple fields)"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (comma-separated for multiple fields)"
          },
          {
            "in": "query",
            "name": "authorId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by author ID"
          },
          {
            "in": "query",
            "name": "parentId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by parent comment ID"
          },
          {
            "in": "query",
            "name": "dateFrom",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter comments created after this date"
          },
          {
            "in": "query",
            "name": "dateTo",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Filter comments created before this date"
          },
          {
            "in": "query",
            "name": "specialSort",
            "schema": {
              "type": "string",
              "enum": [
                "mostLiked",
                "mostDisliked",
                "newest",
                "oldest"
              ]
            },
            "description": "Special sorting options"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "description": "Search term for comment content"
          },
          {
            "in": "query",
            "name": "minLikes",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Minimum number of likes"
          },
          {
            "in": "query",
            "name": "maxLikes",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Maximum number of likes"
          },
          {
            "in": "query",
            "name": "minReplies",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Minimum number of replies"
          },
          {
            "in": "query",
            "name": "maxReplies",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Maximum number of replies"
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comments retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "comments": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CommentWithDetails"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadata"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments/{id}/replies": {
      "get": {
        "summary": "Get replies to a comment",
        "description": "Retrieve paginated replies to a specific comment",
        "tags": [
          "Comments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Parent comment ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "content",
                "likesCount",
                "dislikesCount",
                "repliesCount"
              ]
            },
            "description": "Field to sort by (comma-separated for multiple fields)"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Sort order (comma-separated for multiple fields)"
          }
        ],
        "responses": {
          "200": {
            "description": "Replies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Replies retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "comments": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CommentWithDetails"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/PaginationMetadata"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Parent comment not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}